# planner: improve the performance of FindFieldName