--- a/pkg/expression/simple_rewriter_test.go
+++ b/pkg/expression/simple_rewriter_test.go
@@ -7,6 +7,7 @@ import (
 	"github.com/pingcap/tidb/pkg/parser/ast"
 	"github.com/pingcap/tidb/pkg/parser/model"
 	"github.com/pingcap/tidb/pkg/types"
+	"github.com/stretchr/testify/require"
 )
 
 // Original function for comparison
@@ -47,6 +48,92 @@ func generateTestData(size int) (types.NameSlice, *ast.ColumnName) {
 		Name:   model.NewCIStr("colZ"), // This will be at the end of the slice
 	}
 	return names, astCol
+}
+
+func TestFindFieldName(t *testing.T) {
+	tests := []struct {
+		name     string
+		names    types.NameSlice
+		astCol   *ast.ColumnName
+		expected int
+		err      error
+	}{
+		{
+			name: "Simple match",
+			names: types.NameSlice{
+				{DBName: model.NewCIStr("db"), TblName: model.NewCIStr("tbl"), ColName: model.NewCIStr("col")},
+			},
+			astCol:   &ast.ColumnName{Schema: model.NewCIStr("db"), Table: model.NewCIStr("tbl"), Name: model.NewCIStr("col")},
+			expected: 0,
+		},
+		{
+			name: "Match with empty schema and table",
+			names: types.NameSlice{
+				{DBName: model.NewCIStr("db"), TblName: model.NewCIStr("tbl"), ColName: model.NewCIStr("col")},
+			},
+			astCol:   &ast.ColumnName{Schema: model.NewCIStr(""), Table: model.NewCIStr(""), Name: model.NewCIStr("col")},
+			expected: 0,
+		},
+		{
+			name: "No match",
+			names: types.NameSlice{
+				{DBName: model.NewCIStr("db"), TblName: model.NewCIStr("tbl"), ColName: model.NewCIStr("col1")},
+			},
+			astCol:   &ast.ColumnName{Schema: model.NewCIStr("db"), Table: model.NewCIStr("tbl"), Name: model.NewCIStr("col2")},
+			expected: -1,
+		},
+		{
+			name: "Match with redundant field",
+			names: types.NameSlice{
+				{DBName: model.NewCIStr("db"), TblName: model.NewCIStr("tbl"), ColName: model.NewCIStr("col"), Redundant: true},
+				{DBName: model.NewCIStr("db"), TblName: model.NewCIStr("tbl"), ColName: model.NewCIStr("col")},
+			},
+			astCol:   &ast.ColumnName{Schema: model.NewCIStr("db"), Table: model.NewCIStr("tbl"), Name: model.NewCIStr("col")},
+			expected: 1,
+		},
+		{
+			name: "Non-unique match",
+			names: types.NameSlice{
+				{DBName: model.NewCIStr("db"), TblName: model.NewCIStr("tbl"), ColName: model.NewCIStr("col")},
+				{DBName: model.NewCIStr("db"), TblName: model.NewCIStr("tbl"), ColName: model.NewCIStr("col")},
+			},
+			astCol: &ast.ColumnName{Schema: model.NewCIStr("db"), Table: model.NewCIStr("tbl"), Name: model.NewCIStr("col")},
+			err:    errNonUniq.GenWithStackByArgs("db.tbl.col", "field list"),
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			require := require.New(t)
+
+			// Test FindFieldNameOriginal
+			idxOriginal, errOriginal := FindFieldNameOriginal(tt.names, tt.astCol)
+			if tt.err != nil {
+				require.Error(errOriginal)
+				require.Equal(tt.err.Error(), errOriginal.Error())
+			} else {
+				require.NoError(errOriginal)
+				require.Equal(tt.expected, idxOriginal)
+			}
+
+			// Test FindFieldName
+			idxOptimized, errOptimized := FindFieldName(tt.names, tt.astCol)
+			if tt.err != nil {
+				require.Error(errOptimized)
+				require.Equal(tt.err.Error(), errOptimized.Error())
+			} else {
+				require.NoError(errOptimized)
+				require.Equal(tt.expected, idxOptimized)
+			}
+
+			// Compare results of both functions
+			require.Equal(idxOriginal, idxOptimized)
+			require.Equal(errOriginal != nil, errOptimized != nil)
+			if errOriginal != nil && errOptimized != nil {
+				require.Equal(errOriginal.Error(), errOptimized.Error())
+			}
+		})
+	}
 }
 
 // go test -bench=^BenchmarkFindFieldName$ -run=^$ -tags intest github.com/pingcap/tidb/pkg/expression