<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: close https://github.com/pingcap/tidb/issues/52112

Problem Summary:
This function becomes sluggish with numerous columns. So, I tried to improve it in this pull request.

### What changed and how does it work?

1. Early continue: We check NotExplicitUsable and ColName first, as these are likely to eliminate many candidates quickly.
2. Avoid repeated access: We precompute dbName.L, tblName.L, and colName.L at the beginning.


Benchmark result:

```console
Running tool: /opt/homebrew/bin/go test -benchmem -run=^$ -tags intest -bench ^BenchmarkFindFieldName$ github.com/pingcap/tidb/pkg/expression

goos: darwin
goarch: arm64
pkg: github.com/pingcap/tidb/pkg/expression
BenchmarkFindFieldName/Original-10-8         	10015628	       120.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkFindFieldName/Optimized-10-8        	37335829	        31.81 ns/op	       0 B/op	       0 allocs/op
BenchmarkFindFieldName/Original-100-8        	  559112	      2155 ns/op	     736 B/op	      10 allocs/op
BenchmarkFindFieldName/Optimized-100-8       	  693345	      1720 ns/op	     736 B/op	      10 allocs/op
BenchmarkFindFieldName/Original-200-8        	  540812	      2150 ns/op	     736 B/op	      10 allocs/op
BenchmarkFindFieldName/Optimized-200-8       	  713084	      1725 ns/op	     736 B/op	      10 allocs/op
BenchmarkFindFieldName/Original-1000-8       	  558850	      2159 ns/op	     736 B/op	      10 allocs/op
BenchmarkFindFieldName/Optimized-1000-8      	  712002	      1795 ns/op	     736 B/op	      10 allocs/op
BenchmarkFindFieldName/Original-10000-8      	  558022	      2184 ns/op	     736 B/op	      10 allocs/op
BenchmarkFindFieldName/Optimized-10000-8     	  682408	      1725 ns/op	     736 B/op	      10 allocs/op
PASS
ok  	github.com/pingcap/tidb/pkg/expression	14.262s
```

### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [ ] Integration test
- [ ] Manual test (add detailed scripts or steps below)
- [ ] No need to test
  > - [ ] I checked and no code files have been changed.
  > <!-- Or your custom  "No need to test" reasons -->

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

```release-note
None
```